Move fake into test code
